requirements.txt
asyncio
flask
pika
pyyaml
telethon
mysql-connector-python
pymongo
requests
pytz